[disc] Break out disc types into separate libraries
Put concrete disc types AudioCD, DVD and BluRay into their own libraries.  This
will be needed for the ripper to import disc.py without creating a circular
dependence.
