+++ src/dartt/disc.py
@@ -20,39 +20,64 @@
 
 from abc import ABC, abstractmethod
 import discid
+from pathlib import Path
 import logging
 
-import dartt.device as device
+import dartt.config as config
 import dartt.musicbrainz as mb
+import dartt.device as device
+
+class Track:
+    def __init__(self, Archive: Path):
+        self._Archive = Archive
+
+class AudioTrack(Track):
+    def __init__(self, ArchivePath: str, TrackInfo: mb.TrackInfo):
+        super().__init__(ArchivePath)
+        self._TrackInfo = TrackInfo
+
+    @property
+    def Number(self):
+        return self._TrackInfo.Number
+
+    @property
+    def Title(self):
+        return self._TrackInfo.Title
+
+    @property
+    def Artist(self):
+        return self._TrackInfo.Artist
+
+    def __repr__(self) -> str:
+        return f'{self._Archive}: {self.Number}. {self.Title} - {self.Artist}'
+
+class VideoTrack(Track):
+    def __init__(self, Config):
+        pass
 
 class Disc(ABC):
     def __init__(self, Dev: device.Device):
         self._Device = Dev
 
+    def rip(self, Config):
+        return []
+
 class AudioDisc(Disc):
     def __init__(self, Dev: device.Device):
         super().__init__(Dev)
 
+    @abstractmethod
+    def getTitle(self):
+        pass
+
+    @abstractmethod
+    def getArtists(self):
+        pass
+
+    @abstractmethod
+    def getTrackInfo(self):
+        pass
+
 class VideoDisc(Disc):
-    def __init__(self, Dev: device.Device):
-        super().__init__(Dev)
-
-class AudioCD(AudioDisc):
-    def __init__(self, Dev: device.Device, Musicbrainz: mb.MusicBrainz):
-        super().__init__(Dev)
-        self._Musicbrainz = Musicbrainz
-
-        logging.debug('Reading disc ID')
-        self._DiscIDInfo = discid.read(self._Device.path)
-        logging.debug(f'Done reading disc ID: {self._DiscIDInfo}')
-
-        Info = self._Musicbrainz.getDiscInfo(self._DiscIDInfo)
-        logging.debug(f'Disc info: {Info}')
-
-class DVD(VideoDisc):
-    def __init__(self, Dev: device.Device):
-        super().__init__(Dev)
-
-class BluRay(VideoDisc):
     def __init__(self, Dev: device.Device):
         super().__init__(Dev)
