+++ tests/conftest.py
@@ -20,7 +20,7 @@
 # You should have received a copy of the GNU Affero General Public License along
 # with dartt. If not, see <https://www.gnu.org/licenses/>.
 
-from typing import Callable, Sequence, Type
+from typing import Callable, Dict, Iterable, Sequence, Type
 import pytest
 
 from dartt.config import Config
@@ -366,6 +366,134 @@ def DiscIDFactory(
         return MockDiscID()
 
     return makeDiscID
+
+class MockDevice:
+    def __init__(
+            self,
+            Name: str,
+            Node: str,
+            Type: str
+    ):
+        if Type == 'CD':
+            self._Properties = {
+                'ID_CDROM_MEDIA': '1',
+                'ID_CDROM_MEDIA_CD': '1',
+                'ID_CDROM_MEDIA_SESSION_COUNT': '1',
+                'ID_CDROM_MEDIA_TRACK_COUNT': '21',
+                'ID_CDROM_MEDIA_TRACK_COUNT_AUDIO': '21',
+            }
+        elif Type == 'DVD':
+            self._Properties = {
+                'ID_CDROM_MEDIA': '1',
+                'ID_CDROM_MEDIA_DVD': '1',
+                'ID_CDROM_MEDIA_SESSION_COUNT': '21',
+                'ID_CDROM_MEDIA_STATE': 'complete',
+                'ID_CDROM_MEDIA_TRACK_COUNT': '1',
+                'ID_CDROM_MEDIA_TRACK_COUNT_DATA': '1',
+                'ID_FS_APPLICATION_ID': 'DVDAfterEdit',
+                'ID_FS_BLOCKSIZE': '2048',
+                'ID_FS_LABEL': 'A_GOOD_MOVIE',
+                'ID_FS_LABEL_ENC': 'A_GOOD_MOVIE',
+                'ID_FS_LOGICAL_VALUE_ID': 'A_GOOD_MOVIE',
+                'ID_FS_TYPE': 'udf',
+                'ID_FS_USAGE': 'filesystem',
+                'ID_FS_UUID': '33fe932d00000000',
+                'ID_FS_UUID_ENC': '33fe932d00000000',
+                'ID_FS_VERSION': '1.02',
+                'ID_FS_VOLUME_ID': 'A_GOOD_MOVIE',
+                'ID_FS_VOLUME_SET_ID': '33fe932d',
+            }
+        elif Type == 'BD':
+            self._Properties = {
+                'ID_CDROM_MEDIA': '1',
+                'ID_CDROM_MEDIA_BD': '1',
+                'ID_CDROM_MEDIA_SESSION_COUNT': '21',
+                'ID_CDROM_MEDIA_STATE': 'complete',
+                'ID_CDROM_MEDIA_TRACK_COUNT': '1',
+                'ID_CDROM_MEDIA_TRACK_COUNT_DATA': '1',
+                'ID_FS_APPLICATION_ID': 'APPLICATION_ID',
+                'ID_FS_BLOCKSIZE': '2048',
+                'ID_FS_LABEL': 'Another_Good_Movie',
+                'ID_FS_LABEL_ENC': 'Another_Good_Movie',
+                'ID_FS_LOGICAL_VALUE_ID': 'Another_Good_Movie',
+                'ID_FS_TYPE': 'udf',
+                'ID_FS_USAGE': 'filesystem',
+                'ID_FS_UUID': '33fe932d91fa3de2',
+                'ID_FS_UUID_ENC': '33fe932d91fa3de2',
+                'ID_FS_VERSION': '2.50',
+                'ID_FS_VOLUME_ID': 'Another_Good_Movie',
+                'ID_FS_VOLUME_SET_ID': '33FE932D91fA3DE2_VOLUME_SET_ID',
+            }
+        else:
+            self._Properties = {}
+
+        self._Properties['ID_CDROM'] = '1'
+        self._SysName = Name
+        self._DeviceNode = Node
+
+    @property
+    def properties(
+            self
+    ) -> Dict[str, str]:
+        return self._Properties
+
+    def keys(
+            self
+    ) -> Iterable[str]:
+        return self._Properties.keys()
+
+    @property
+    def sys_name(
+            self
+    ) -> str:
+        return self._SysName
+
+    @property
+    def device_node(
+            self
+    ) ->str:
+        return self._DeviceNode
+
+@pytest.fixture
+def deviceFactory(
+        request
+) -> Callable[[Sequence[str]], MockDevice]:
+    def makeDevice(
+            Name: str,
+            Node: str,
+            DiscType: str
+    ) -> MockDevice:
+        return MockDevice(Name, Node, DiscType)
+
+    return makeDevice
+
+class MockDevices:
+    def __init__(
+            self,
+            Names: Sequence[str],
+            Nodes: Sequence[str],
+            Type: str
+    ):
+        self._Names = Names
+        self._Nodes = Nodes
+        self._Type = Type
+
+    def match(self, **kwargs):
+        return [ MockDevice(Name, Node, self._Type)
+                 for Name, Node in zip(self._Names, self._Nodes) ]
+
+@pytest.fixture
+def devicesFactory(
+        request
+) -> Callable[[Sequence[str]], MockDevices]:
+    def makeDevices(
+            Names: Sequence[str],
+            Nodes: Sequence[str],
+            DiscType: str
+    ) -> MockDevices:
+        return MockDevices(Names, Nodes, DiscType)
+
+    return makeDevices
 
 @pytest.fixture
 def commandFactory(
